feat: :sparkles: Aplicando controlle de fluxo para requisições inválidas

+++ app_locadora_carros/app/Http/Controllers/MarcaController.php
@@ -53,6 +53,9 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -80,6 +83,9 @@ class MarcaController extends Controller
         //echo "<hr>";
         //print_r($marca->getAttributes()); // dados antigos
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -94,6 +100,9 @@ class MarcaController extends Controller
     {
 
         $marca= $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso'];
     }
